Update README.md
This commit updates the README.md file to provide comprehensive details about the Preloved With Friend App. It includes descriptions of key features such as the user-friendly interface, advanced search functionality, seller dashboard, and secure payment gateway. 

Additionally, the update highlights the use of Figma for designing the user interface and Wix as the development platform, ensuring clarity on the technologies utilized in this project. This information aims to give contributors and users a better understanding of the app's purpose and functionality.